moving seldom used functions
