+++ cxgo/actions/literals.go
@@ -3,9 +3,8 @@ package actions
 import (
 	"github.com/skycoin/cx/cx/ast"
 	"github.com/skycoin/cx/cx/constants"
+	"github.com/skycoin/cx/cx/globals"
 	"github.com/skycoin/skycoin/src/cipher/encoder"
-
-	"github.com/skycoin/cx/cx"
 )
 
 // SliceLiteralExpression handles literal expressions by converting it to a series of `append` expressions.
@@ -17,7 +16,7 @@ func SliceLiteralExpression(typSpec int, exprs []*ast.CXExpression) []*ast.CXExp
 		panic(err)
 	}
 
-	symName := cxcore.MakeGenSym(constants.LOCAL_PREFIX)
+	symName := globals.MakeGenSym(constants.LOCAL_PREFIX)
 
 	// adding the declaration
 	slcVarExpr := ast.MakeExpression(nil, CurrentFile, LineNo)
@@ -58,7 +57,7 @@ func SliceLiteralExpression(typSpec int, exprs []*ast.CXExpression) []*ast.CXExp
 			} else {
 				// We need to create a temporary variable to hold the result of the
 				// nested expressions. Then use that variable as part of the slice literal.
-				out := ast.MakeArgument(cxcore.MakeGenSym(constants.LOCAL_PREFIX), expr.FileName, expr.FileLine)
+				out := ast.MakeArgument(globals.MakeGenSym(constants.LOCAL_PREFIX), expr.FileName, expr.FileLine)
 				outArg := getOutputType(expr)
 				out.AddType(constants.TypeNames[outArg.Type])
 				out.CustomType = outArg.CustomType
@@ -87,7 +86,7 @@ func SliceLiteralExpression(typSpec int, exprs []*ast.CXExpression) []*ast.CXExp
 		expr.IsArrayLiteral = false
 	}
 
-	symNameOutput := cxcore.MakeGenSym(constants.LOCAL_PREFIX)
+	symNameOutput := globals.MakeGenSym(constants.LOCAL_PREFIX)
 
 	symOutput := ast.MakeArgument(symNameOutput, CurrentFile, LineNo).AddType(constants.TypeNames[typSpec])
 	symOutput.IsSlice = true
@@ -193,7 +192,7 @@ func ArrayLiteralExpression(arrSizes []int, typSpec int, exprs []*ast.CXExpressi
 		panic(err)
 	}
 
-	symName := cxcore.MakeGenSym(constants.LOCAL_PREFIX)
+	symName := globals.MakeGenSym(constants.LOCAL_PREFIX)
 
 	arrVarExpr := ast.MakeExpression(nil, CurrentFile, LineNo)
 	arrVarExpr.Package = pkg
@@ -252,7 +251,7 @@ func ArrayLiteralExpression(arrSizes []int, typSpec int, exprs []*ast.CXExpressi
 		}
 	}
 
-	symNameOutput := cxcore.MakeGenSym(constants.LOCAL_PREFIX)
+	symNameOutput := globals.MakeGenSym(constants.LOCAL_PREFIX)
 
 	symOutput := ast.MakeArgument(symNameOutput, CurrentFile, LineNo).AddType(constants.TypeNames[typSpec])
 	// symOutput.Lengths = append(symOutput.Lengths, arrSizes[len(arrSizes)-1])
